Add max request size of 100 MB.
This is on the first line of the server function, and was
implemented for the upcoming web version to prevent abuse/accidents.
A user could comment this out in their installed version to analyze
larger files or gene set databases (we'll come up with a more elegant
solution one of these days).