Accept Sample Info table as CSV file
NanoTube now asks for a CSV file for the sample info table. It then
provides the user with a list of columns of the table, from which the
user can specify the column containing comparison group identifiers.
The user can then select the "base group", or the group against
which other groups are compared in differential expression analysis.